+++ packages/apollo-link-schema/src/__tests__/schemaLink.ts
@@ -63,7 +63,7 @@ describe('SchemaLink', () => {
   });
 
   it('throws if no arguments given', () => {
-    expect(() => new SchemaLink()).toThrow();
+    expect(() => new (SchemaLink as any)()).toThrow();
   });
 
   it('correctly receives the constructor arguments', () => {
@@ -135,9 +135,10 @@ describe('SchemaLink', () => {
     );
   });
 
-  it('passes operation context into execute', done => {
+  it('passes operation context into execute with context function', done => {
     const next = jest.fn();
     const contextValue = { some: 'value' };
+    const contextProvider = jest.fn(operation => operation.getContext());
     const resolvers = {
       Query: {
         sampleQuery: (root, args, context) => {
@@ -153,7 +154,10 @@ describe('SchemaLink', () => {
       typeDefs,
       resolvers,
     });
-    const link = new SchemaLink({ schema: schemaWithResolvers });
+    const link = new SchemaLink({
+      schema: schemaWithResolvers,
+      context: contextProvider,
+    });
     const observable = execute(link, {
       query: sampleQuery,
       context: contextValue,
@@ -162,8 +166,55 @@ describe('SchemaLink', () => {
       next,
       error => done.fail("Shouldn't call onError"),
       () => {
-        expect(next).toHaveBeenCalledTimes(1);
-        done();
+        try {
+          expect(next).toHaveBeenCalledTimes(1);
+          expect(contextProvider).toHaveBeenCalledTimes(1);
+          done();
+        } catch (e) {
+          done.fail(e);
+        }
+      },
+    );
+  });
+
+  it('passes static context into execute', done => {
+    const next = jest.fn();
+    const contextValue = { some: 'value' };
+    const resolver = jest.fn((root, args, context) => {
+      try {
+        expect(context).toEqual(contextValue);
+      } catch (error) {
+        done.fail('Should pass context into resolver');
+      }
+    });
+
+    const resolvers = {
+      Query: {
+        sampleQuery: resolver,
+      },
+    };
+    const schemaWithResolvers = makeExecutableSchema({
+      typeDefs,
+      resolvers,
+    });
+    const link = new SchemaLink({
+      schema: schemaWithResolvers,
+      context: contextValue,
+    });
+    const observable = execute(link, {
+      query: sampleQuery,
+    });
+    observable.subscribe(
+      next,
+      error => done.fail("Shouldn't call onError"),
+      () => {
+        try {
+          expect(next).toHaveBeenCalledTimes(1);
+          expect(resolver).toHaveBeenCalledTimes(1);
+          done();
+        } catch (e) {
+          done.fail(e);
+        }
       },
     );
   });